Start of actual business-level security implementation

+++ spring-boot-zero-to-hero/src/main/groovy/com/thirdstart/spring/zerotohero/services/ContactService.groovy
@@ -1,12 +1,13 @@
 package com.thirdstart.spring.zerotohero.services
 
 import com.thirdstart.spring.zerotohero.domain.Contact
+import com.thirdstart.spring.zerotohero.domain.User
 import com.thirdstart.spring.zerotohero.repositories.ContactRepository
 import com.thirdstart.spring.zerotohero.util.spring.SimpleValidator
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.security.access.AccessDeniedException
 import org.springframework.security.access.annotation.Secured
 import org.springframework.security.access.prepost.PreAuthorize
-import org.springframework.security.core.userdetails.User
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
 
@@ -27,11 +28,19 @@ class ContactService {
     @Autowired
     UserService userService
 
+    User getCurrentUser() {
+        return userService.currentUser
+    }
+
     Contact findOne(Long id) {
         Contact contact = contactRepository.findOne(id)
 
         if ( !contact ) {
             throw new EntityNotFoundException("No contact found for id ${id}.")
+        }
+
+        if ( !contact.canBeAccessedBy( currentUser ) ) {
+            throw new AccessDeniedException('Access denied.')
         }
 
         return contact
@@ -44,6 +53,12 @@ class ContactService {
             throw new Exception("We don't save Nope! around here!")
         }
 
+        if ( contact.id && contact.canBeAccessedBy( currentUser ) ) {
+            throw new AccessDeniedException("Access denied.")
+        } else if ( !contact.id ) {
+            contact.createdBy = userService.currentUser
+        }
+
         simpleValidator.validateAndThrowOnFailure( contact )
 
         contactRepository.save(contact)
@@ -52,6 +67,11 @@ class ContactService {
     }
 
     void delete(Contact contact) {
+
+        if ( !contact.canBeAccessedBy( currentUser ) ) {
+            throw new AccessDeniedException("Access denied.")
+        }
+
         contactRepository.delete(contact)
     }
